idl/ast/rpc: Add a member for storing enum string
Dispatch loop will call functions based on function ID it receives,
function ID's are generated from Rpc names. These names go as enums into
dispatch loop's switch..case for identifying which function needs to be
called. As this needs to be generated twice (both header and dispatch
loop), store it as a member.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>