idl/ast: Fix endless loop, Resolve function args
In situations like this
----
projection <struct bar> bar {
  rpc int (*baz)(projection foo *foo);
}

projection <struct foo> foo {
  projection bar *bar;
}
---
there is a circular dependency. During copy types, this causes an
infinite recursion if we try to resolve the function arguments. So,
let's just disable to fix this. Most of the time, the type information
of all the projection fields is not necessary for 'baz' to operate.
Disable this for now using if (0).

Also resolve function arguments if it is marked unresolved by our grammar
parser.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>