idl/ast: Introduce modify specs pass
If a parameter is marked bind, the caller should rather marshall
"other_ref" cptr. So far, during creation of containers, "my_ref" is
always marked as "in".

Introduce a modify_specs to fix the specs {alloc,bind,dealloc} in such a
way that, if bind is specified, "other_ref" would be marked as "in".

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>