ast: Fix unnecessary re-marshalling of a variable
Before this fix, the compiler was unnecessarily re-marshalling projection
fields. This happened because when our marshalling technique marshals a
projection type, it also marshals the fields of the projection type.
Thereafter, the technique also remarshals these fields separately, as these
fields are stored as parameters of the rpc. The side-effect of this -- we were
unnecessarily run out of registers. The fix -- marshal a variable only if it
does not have any marshalling info (a "Marshal_type")