code_gen: Fix wrong additin of container statemnts
The logic is to declare and allocate containers **only for those parameters of
an rpc that have containers**. This fix moves the container addition code for a
function pointer parameter inside the "if p has a container" condition block.
I am pretty sure this error was unintentional by the author -- But it caused
the compiler to crash at line no. 213, while trying to access the container of
a parameter, which does not have a container.