ast: Add spec propagation & bind inference support
Add specifier propagation among projections, and nested projections. Bind
inference is also done in the technique. For functions and function pointers in
the IDL, the algo assumes function pointers are always called from the kernel,
and normal function calls are called from the isolated domain. The main
implementation is in astprint.cpp. File should be renamed.