code_gen: Fix container_of's arg, and add fp check
+ The container_of macro's 2nd arg, the type arg, is named as
<real_type_of_projection>_container instead of <projection_name>_container.

+ The fp check fix: allocate and link variables on the caller side (the
isolated domain) only if the variable's alloc(caller) specifer is true AND if
the variable is not a parameter or return variable of a function pointer. This
is because we assume function pointers would be called from the kernel only.
Although this may be true in most cases in the linux environment, it may not
always be the case. This function pointer issue needs to be handled in a more
generic way (an important TODO.)